# Optimization in llama.py