No need to slice A and B twice in LLaMa complex multiplication

## Before:
```
test/external/external_test_speed_llama.py::TestLLaMASpeed::test_llama_compile using <class 'tinygrad.runtime.ops_metal.MetalCodegen'>
testing llama python run time
built model
assigned empty tensors, doing warmup
codegen         mean runtime:  112.46ms, runs:   315.72,   82.73,   83.22,   84.40,   83.35,  121.45,   86.78,   90.19,   88.71,   88.05
methodcache     mean runtime:   88.56ms, runs:    75.47,  119.04,   81.63,   81.04,   79.23,   83.41,   82.42,   79.36,  122.67,   81.34
profile         mean runtime:  455.47ms, runs:   529.99,  515.67,  444.73,  426.64,  428.02,  471.64,  431.93,  435.21,  436.11,  434.71
```

## After:

```
test/external/external_test_speed_llama.py::TestLLaMASpeed::test_llama_compile using <class 'tinygrad.runtime.ops_metal.MetalCodegen'>
testing llama python run time
built model
assigned empty tensors, doing warmup
codegen         mean runtime:   91.79ms, runs:   209.08,   73.39,   73.41,   73.74,   73.85,   76.95,   74.96,   75.89,  110.68,   75.98
methodcache     mean runtime:   73.84ms, runs:    68.23,   70.41,   70.17,   70.56,   70.48,   70.47,   70.44,   70.49,  106.12,   71.01
profile         mean runtime:  435.62ms, runs:   493.45,  410.60,  418.22,  420.68,  421.02,  439.58,  460.59,  431.14,  431.27,  429.67
```